task2/requirements.txt
alabaster==0.7.12
appdirs @ file:///home/conda/feedstock_root/build_artifacts/appdirs_1603108395799/work
astor==0.8.1
atomicwrites==1.4.0
audioread @ file:///home/conda/feedstock_root/build_artifacts/audioread_1607457396158/work
autoflake==1.4
Babel==2.9.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bandit==1.7.0
black==20.8b1
bleach==3.3.0
brotlipy==0.7.0
bumpversion==0.5.3
certifi==2020.12.5
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1613413861439/work
chardet @ file:///home/conda/feedstock_root/build_artifacts/chardet_1610093490430/work
click==7.1.2
coverage==5.5
coveralls==1.7.0
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1616851476134/work
cycler==0.10.0
darglint==1.8.0
decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
distlib==0.3.1
doc8==0.8.0
docopt==0.6.2
docutils==0.17
efficientnet-pytorch==0.7.1
einops==0.3.0
eradicate==1.0
filelock==3.0.12
fire @ file:///home/conda/feedstock_root/build_artifacts/fire_1611345183397/work
flake8-2020==1.6.0
flake8-bandit==2.1.2
flake8-broken-line==0.2.1
flake8-bugbear==19.8.0
flake8-builtins==1.5.3
flake8-commas==2.0.0
flake8-comprehensions==3.4.0
flake8-debugger==3.2.1
flake8-docstrings==1.6.0
flake8-fixme==1.1.1
flake8-isort==3.0.1
flake8-logging-format==0.6.0
flake8-mutable==1.2.0
flake8-polyfill==1.0.2
flake8-quotes==2.1.2
flake8-rst-docstrings==0.0.12
flake8-string-format==0.2.3
flake8-variables-names==0.0.3
gitdb==4.0.7
GitPython==3.1.14
h5py @ file:///home/conda/feedstock_root/build_artifacts/h5py_1617647518676/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
imagesize==1.2.0
ipython @ file:///tmp/build/80754af9/ipython_1617120885885/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.3
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1604322295622/work
librosa @ file:///home/conda/feedstock_root/build_artifacts/librosa_1622025601369/work
llvmlite==0.36.0
loguru @ file:///home/conda/feedstock_root/build_artifacts/loguru_1618589835436/work
MarkupSafe==1.1.1
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-base_1594091694890/work
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1618853849286/work
mkl-service==2.3.0
more-itertools==8.7.0
mypy-extensions==0.4.3
numba @ file:///home/conda/feedstock_root/build_artifacts/numba_1616707886102/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1620830962040/work
olefile==0.46
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1612459636436/work
pandas==1.1.4
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pathspec==0.8.1
pbr==5.5.1
pep8-naming==0.9.1
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///tmp/build/80754af9/pillow_1617383569452/work
pkginfo==1.7.0
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1606467285986/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1616415428029/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pydocstyle==5.1.1
Pygments==2.8.1
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing==2.4.7
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1610291447907/work
pytest==6.2.3
pytest-cov==2.11.1
pytest-runner==4.4
python-dateutil==2.8.1
python-dev-tools==2020.9.10
pytorch-ignite==0.4.4
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1612179539967/work
pyupgrade==2.11.0
PyYAML==5.4.1
readme-renderer==29.0
regex==2021.4.4
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1608156231189/work
requests-toolbelt==0.9.1
resampy==0.2.2
restructuredtext-lint==1.3.2
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1621370412049/work
scipy @ file:///tmp/build/80754af9/scipy_1618855647378/work
six @ file:///tmp/build/80754af9/six_1605205327372/work
smmap==4.0.0
snowballstemmer==2.1.0
SoundFile @ file:///home/conda/feedstock_root/build_artifacts/pysoundfile_1607011569782/work
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
stevedore==3.3.0
termcolor==1.1.0
testfixtures==6.17.1
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tokenize-rt==4.1.0
torch==1.8.1
torch-audiomentations==0.7.0
torchaudio==0.8.0a0+e4e171a
torchvision==0.9.1
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1604105045397/work
tox==3.23.0
tox-travis==0.12
tqdm @ file:///tmp/build/80754af9/tqdm_1615925068909/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
twine==1.13.0
typed-ast==1.4.2
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1622056799390/work
virtualenv==20.4.3
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
wemake-python-styleguide==0.14.1
